Add support for storing blocks in compressed form using deflate compression, enabled by the `--compress' command line flag.
Alter the semantics of the read_block() and write_block() API functions
so zero block detection and MD5 computation are performed by the lowest
layer.  This is required because the MD5 in the Content-MD5 and If-Match
HTTP headers applies to the compressed content.

Since auto-detection of block size no longer works if the first block
is compressed, start including a "x-amz-meta-s3backer-blocksize" header
with the first block as well. Include some backward compatibility code
to fall back to using the Content-Length (as we were doing before),
assuming the block is not compressed, if the new header is not found.